pyproject.toml
[project]
name = "inkcre-core"
version = "0.1.0"
description = "Backend of InKCre."
authors = [
    {name = "Lan_zhijiang", email = "lanzhijiang@foxmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiohttp (>=3.12.14,<4.0.0)",
    "tencentcloud-sdk-python-lke (>=3.0.1427,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "openai (>=1.97.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "pgvector (>=0.4.1,<0.5.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
